auth_project/requirements.txt
django>=4.2
djangorestframework>=3.14
bcrypt>=4.0
PyJWT>=2.8
psycopg2-binary>=2.9
